Feat/external code editor (#255)
* wip: external code editor

* improve code editor detection and fix hotkeys

* fixes to shortcuts

* try fix opening .app bundles on macOS

* improve jetbrains IDE detection on windows

---------

+++ crates/gitcomet-tree-sitter-alloc/src/lib.rs
@@ -276,7 +276,7 @@ unsafe fn usable_size(ptr: *mut c_void) -> usize {
 
     #[cfg(windows)]
     {
-        unsafe { _msize(ptr) as usize }
+        unsafe { _msize(ptr) }
     }
 
     #[cfg(not(any(
